+++ src/main/java/com/example/practiceDemo1/App.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 /**
  * Hello world!
