+++ src/main/java/com/example/practiceDemo1/ArrayDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class ArrayDemo {
     public static void main(String[] args) {
