+++ src/main/java/com/example/practiceDemo1/ArrayListDemo.java
@@ -1,11 +1,10 @@
-package com.example;
+package com.example.practiceDemo1;
 
 import java.util.ArrayList;
-import java.util.List;
 
-public class ListDemo {
+public class ArrayListDemo {
     public static void main(String[] args) {
-        List<String> animal=new ArrayList<>();
+        ArrayList<String> animal=new ArrayList<>();
         animal.add("cat");
         animal.add("dog");
         animal.add("cow");
