+++ src/main/java/com/example/practiceDemo1/ConstructorDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class ConstructorDemo {
     public static void main(String[] args) {
