+++ src/main/java/com/example/practiceDemo1/DoWhileDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class DoWhileDemo {
     public static void main(String[] args) {
