+++ src/main/java/com/example/practiceDemo1/InheritanceDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 class Vehicle{
     public void Sound(){
