+++ src/main/java/com/example/practiceDemo1/IteratorDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 import java.util.ArrayList;
 import java.util.Iterator;
