+++ src/main/java/com/example/practiceDemo1/P1.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class P1 {
     public static void main(String[] args) {
