+++ src/main/java/com/example/practiceDemo1/P3.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 import java.util.ArrayList;
 import java.util.HashSet;
