+++ src/main/java/com/example/practiceDemo1/ScannerDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 import java.util.Scanner;
