+++ src/main/java/com/example/practiceDemo1/ThreadImplementDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class ThreadImplementDemo implements Runnable{
     public static void main(String[] args) {
