+++ src/main/java/com/example/practiceDemo1/ThreadOrderingDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class ThreadOrderingDemo extends Thread{
     public static void main(String[] args) {
