+++ src/main/java/com/example/practiceDemo1/WhileDemo.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.practiceDemo1;
 
 public class WhileDemo {
     public static void main(String[] args) {
